BUGFIX: The Height and Width fields in Fixed Dimensions blocks no longer get stuck in readonly mode.
The issue was that the code we added to ensure that newly-generated blocks also
got the readOnly toggle events applied to them was redundant. Wagtail's
StreamField system already calls the fixed_dimentions() function each time a new
DimensionsOptionsBlock is instantiated in the form.

I also streamlined the javascript a bit for simplicity, and added comments.