Updated the RelatedLinksBlock to use a ListBlock instead of a StreamBlock.
This offers a much more intuitive interface to the user, and makes much more
sense from a programming perspective as well. I believe this block, and
CaptionedImageCarouselBlock, which both used StreamBlocks in this way, were
written before we learned about ListBlock.

I'll be merging CaptionedImageCarouselBlock into ImageCaraouselBlock (which does
use ListBlock) soon, which will remove the last misuse of StreamBlock.