Removed CaptionedImageCarouselBlock. Added ImageGalleryBlock.
ImageGalleryBlock can be either a regular image gallery, with all thei mages
displayed all at once. Or it can be an image slider, showing only one images at
a time with thumbnails below to act as controls on which image to display.

It's pretty much done, but I'm sure some improvements can be made to how it
interacts with its parent's size.