BUGFIX: VideoBlocks no longer require two clicks to play on mobile devices.
My oh-so-clever solution for getting the video to play when the user clicks the
title overlay block turned out to only worked on desktop. On mobile, I had to do
some  shenanigans with unclickability rules to ensure that the initial tap went
through directly to the iframe. I also added some code to make the title div
vanish when tapped, which is needed on iPads since they don't play the video in
full screen.

The new utils.js file currently only adds a simplified Modernizr test to check
if the user is on a mobile device.