BUGFIX: Uses of block.id in templates now have a backup for when it's not available.
For unknown reasons, the "block" object is a different class of object when it's
rendered at the top level of a StreamField, as opposed to inside a column layout.
This causes the block.id attribute to be missing, so I added a template tag
to generate a unique ID for those situations. This ensures that if, say, multiple
ImageGalleryPanel sliders are on the same page at the top level, their controls
will work properly.