BUGFIX: Column layouts no longer allow left-right scrolling on portrait tablets.
There's some sort of strangeness with the way that Bootstrap's column CSS rules
work. Perhaps we're not applying all the right CSS? Perhaps there's a bootstrap
class that we forgot to include alongside row?

We ended up finding a fairly elegant colution, but we're still pretty baffled as
to why this heavily-used layout has this apparently gigantic bug.