UN-BUGFIX: Removed some CSS that was making too-small images look semi-OK on carousels.
We had previously thought it would make sense to use CSS to scale up small images
to fit with the rest of the Carousel content. But that's slightly unreliable,
because of the vagaries of CSS resizing. So we decided to remove that "fix" and
instead just insist that users provide images that are large enough for any grid
element into which they get placed, so that Wagtail can properly scale/crop them.