FEATURE: Image caption/credit can now have links, sub/superscript, italics, and bold.
I implemented this by creating the custom_bleach template filter, which is the
same as the usual bleach template filter, but lets you specify a different set
of allowed tags from the base set defined in settings.py. This lets specific
fields have certain kinds of tags that other fields are not allowed to have.